+++ cde_analyzer/actions/strip_phrases/cli.py
@@ -11,31 +11,31 @@ def register_subparser(subparser: ArgumentParser):
     #     help="Remove curated phrases from specific paths in a JSON document.",
     # )
     subparser.add_argument(
-        "-i", "--input", required=True, help="Path to input JSON file."
+        "--input", "-i", required=True, help="Path to input JSON file."
     )
     subparser.add_argument(
-        "-m",
         "--model",
+        "-m",
         choices=MODEL_REGISTRY.keys(),
         required=True,
         help="Top-level Pydantic model name for parsing the input JSON.",
     )
     subparser.add_argument(
-        "-p",
         "--phrases",
+        "-p",
         required=True,
         help="Path to phrases file (JSON, CSV, or TSV).",
     )
     subparser.add_argument(
-        "-o", "--output", required=True, help="Path to output JSON file."
+        "--output", "-o", required=True, help="Path to output JSON file."
     )
     # subparser.add_argument(
     #     "-t", "--tids", required=True, help="Path to JSON file with list of tids."
     # )
     # This should be moved to post-processing. Inefficient and memory hungry
     subparser.add_argument(
-        "-d",
         "--diff",
+        "-d",
         action="store_true",
         help="Show diff between original and cleaned JSON",
     )
@@ -43,14 +43,14 @@ def register_subparser(subparser: ArgumentParser):
         "--diff-output", type=str, help="Path to file for writing diff information."
     )
     subparser.add_argument(
-        "-c", "--color", action="store_true", help="Colorize diff output."
+        "--color", "-c", action="store_true", help="Colorize diff output."
     )
     subparser.add_argument(
         "--summary", action="store_true", help="Show a summary of lines changed lines."
     )
     subparser.add_argument(
-        "-C",
         "--context",
+        "-C",
         type=int,
         default=3,
         help="Number of context lines before and after changes.",