+++ cde_analyzer/utils/__init__.py
@@ -11,6 +11,15 @@ from .logger import log_if_verbose, logging
 from .analyzer_state import get_verbosity, set_verbosity
 from .helpers import safe_nested_increment, safe_nested_append
 from .constants import MODEL_REGISTRY
+from .cli_args import (
+    add_input_output_args,
+    add_verbosity_args,
+    add_model_arg,
+    add_field_args,
+    add_match_args,
+    add_pretty_print_args,
+    add_dry_run_arg,
+)
 
 __all__ = [
     "log_if_verbose",
@@ -20,4 +29,11 @@ __all__ = [
     "safe_nested_increment",
     "safe_nested_append",
     "MODEL_REGISTRY",
+    "add_input_output_args",
+    "add_verbosity_args",
+    "add_model_arg",
+    "add_field_args",
+    "add_match_args",
+    "add_pretty_print_args",
+    "add_dry_run_arg",
 ]